% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesson_loader.R
\name{empty_lessons_result}
\alias{empty_lessons_result}
\title{Create an empty lessons result data.frame}
\usage{
empty_lessons_result()
}
\value{
A zero-row data.frame with columns \code{lesson_id}, \code{package}, \code{description}
}
\description{
Create an empty lessons result data.frame
}
\keyword{internal}
