% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-default operator}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{Left-hand side value}

\item{rhs}{Right-hand side default value}
}
\value{
\code{lhs} if non-NULL, otherwise \code{rhs}
}
\description{
Returns \code{lhs} if non-NULL, otherwise \code{rhs}. Provides compatibility
with R versions prior to 4.4 which lack a built-in \verb{\%||\%} operator.
}
\keyword{internal}
