% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_discovery.R
\name{invalidate_lesson_cache}
\alias{invalidate_lesson_cache}
\title{Clear the lesson discovery cache}
\usage{
invalidate_lesson_cache()
}
\value{
Invisible NULL
}
\description{
Clears the session-level cache of discovered lesson packages and lesson
index. Call this after installing or removing packages that contain
blendtutor lessons so that they are re-discovered.
}
