% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educator_tools.R
\name{lesson_yaml_template}
\alias{lesson_yaml_template}
\title{Generate a lesson YAML template string}
\usage{
lesson_yaml_template(lesson_name, title = NULL)
}
\arguments{
\item{lesson_name}{Lesson file name stem}

\item{title}{Display title (defaults to humanised lesson_name)}
}
\value{
Character string of YAML content
}
\description{
Generate a lesson YAML template string
}
\keyword{internal}
