% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educator_tools.R
\name{print_package_summary}
\alias{print_package_summary}
\title{Print a summary of the created lesson package}
\usage{
print_package_summary(path, lesson_name)
}
\arguments{
\item{path}{Character string path to the created package}

\item{lesson_name}{Character string name of the example lesson}
}
\value{
Invisible NULL
}
\description{
Print a summary of the created lesson package
}
\keyword{internal}
