% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_operations.R
\name{retrieve_student_code}
\alias{retrieve_student_code}
\title{Retrieve student code from string or file}
\usage{
retrieve_student_code(code_string, code_file)
}
\arguments{
\item{code_string}{Optional character string of code}

\item{code_file}{Path to temp file containing code}
}
\value{
Character string containing the student's code
}
\description{
Gets code either from direct string argument or by reading the temp file.
If code_string is provided, also saves it to the temp file for persistence.
}
\keyword{internal}
